pyproject.toml
[tool.poetry]
name = "Thrower"
version = "0.1.0"
description = ""
authors = ["Baranov_DR <mrramzes37@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.1"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
cachetools = "^5.3.1"
pillow = "^10.0.0"
httpx = "^0.25.0"
jinja2 = "^3.1.2"
table2ascii = "^1.1.1"
aiosqlite = "^0.19.0"
pydantic-settings = "^2.0.3"
asyncmy = "^0.2.8"

[tool.poetry.group.music.dependencies]
pynacl = "^1.5.0"
lavalink = "^4.0.6"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.23.2"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
